# openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Petstore
  description: A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
  termsOfService: https://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: https://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://petstore.swagger.io/api
paths:
...
  /owners/{id}:
    get:
      summary: Returns a owner by ID
      description: Returns a owner based on a single ID
      operationId: findOwnerByID
      parameters:
        - name: id
          in: path
          description: ID of owners to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: owners response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletes a owner by ID
      description: deletes a single owner based on the ID supplied
      operationId: deleteOwner
      parameters:
        - name: id
          in: path
          description: ID of owner to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: owner deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Owner:
      allOf:
        - $ref: '#/components/schemas/NewOwner'
        - required:
            - id
          properties:
            id:
              type: integer
              format: int64
              description: Unique id of the owner
    NewOwner:
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the owner
        tag:
          type: string
          description: Type of the owner
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message